Update lib.rs
I have simplified the code by combining functions that appear to be performed the same task.

For example, the create_whoami and inner_create_whoami functions both perform the same task of creating a new profile and linking it to the "agents" path and the agent's pubkey. These two functions have now been combined into one function to reduce redundancy.

Additionally, the use of the get_peers function in both create_whoami and inner_create_whoami have been refactored to a single function that is called by both, rather than duplicating the function call.

The create_imported_profile function also performs similar tasks to the other two functions, so I have taken a chance and combined it with them as well.
Haven't tested yet but should work.. give it a try.